ice_breaker/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
distlib==0.3.6
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.12.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
install==1.3.5
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
langsmith==0.0.87
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==1.12.0
openapi-schema-pydantic==1.2.4
packaging==23.2
pathspec==0.11.1
pipenv==2023.5.19
platformdirs==3.5.1
pydantic==1.10.8
python-dotenv==1.0.1
PyYAML==6.0
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tiktoken==0.5.2
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.9.0
typing-inspect==0.9.0
urllib3==2.0.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
yarl==1.9.2
